Project_03_01 started and mostly completed.
Throughout the course of the day, I added in many features into the "ph_styles.css" that include, but are not limited to adding properties when hovering over and clicking hyperlinks, margin and indent spacings and many font changes that include colors, families among other things. 

Comments and general debugging will be needed to fully complete this project, but I have gotten through all 18 questions laid out so far.