Project_03_01 Finished.
After an extensive session of debugging, I was finally able to make the website appear as it does within the instructions. I figured this out by changing my formatting with class and ID selectors as well as making sure padding and margins were't separated by commas (which changed a lot).
.
Comments were also added in addition to this and have been completed with at least at what I believe to be every major new area of code.